# Conflicts:
#	ensysmod/schemas/energy_model.py

+++ ensysmod/model/energy_source.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, DECIMAL
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from ensysmod.database.base_class import Base
@@ -13,7 +13,7 @@ class EnergySource(Base):
     ref_component = Column(Integer, ForeignKey("energy_component.id"), index=True, nullable=False, primary_key=True)
     ref_commodity = Column(Integer, ForeignKey("energy_commodity.id"), index=True, nullable=False)
 
-    commodity_cost = Column(DECIMAL, nullable=True)
+    commodity_cost = Column(Float, nullable=True)
 
     # Relationships
     component = relationship("EnergyComponent")
